Move utils package
