+++ internal/docker/docker.go
@@ -26,10 +26,10 @@ import (
 	"github.com/phayes/freeport"
 	"github.com/rs/zerolog/log"
 
-	"github.com/saucelabs/saucectl/cli/utils"
 	"github.com/saucelabs/saucectl/internal/config"
 	"github.com/saucelabs/saucectl/internal/credentials"
 	"github.com/saucelabs/saucectl/internal/streams"
+	"github.com/saucelabs/saucectl/internal/utils"
 )
 
 var (
