+++ internal/runner/runner.go
@@ -3,9 +3,9 @@ package runner
 import (
 	"context"
 	"github.com/saucelabs/saucectl/cli/command"
-	"github.com/saucelabs/saucectl/cli/utils"
 	"github.com/saucelabs/saucectl/internal/config"
 	"github.com/saucelabs/saucectl/internal/fleet"
+	"github.com/saucelabs/saucectl/internal/utils"
 )
 
 const logDir = "/var/log/cont"
